Added Header to Scores (still need to fix CSS on reduced Browser)
